WIP - cherry picking the OCI layout changes for PoC

+++ acceptance/exporter_test.go
@@ -69,7 +69,9 @@ func testExporterFunc(platformAPI string) func(t *testing.T, when spec.G, it spe
 		var exportedImageName string
 
 		it.After(func() {
-			_, _, _ = h.RunE(exec.Command("docker", "rmi", exportedImageName)) // #nosec G204
+			if !strings.Contains(exportedImageName, "oci") {
+				_, _, _ = h.RunE(exec.Command("docker", "rmi", exportedImageName)) // #nosec G204
+			}
 		})
 
 		when("daemon case", func() {
@@ -255,6 +257,30 @@ func testExporterFunc(platformAPI string) func(t *testing.T, when spec.G, it spe
 						})
 					})
 				})
+			})
+		})
+
+		when("layout case", func() {
+			it.Focus("exports image", func() {
+				exportFlags := []string{"-layout"}
+				if api.MustParse(platformAPI).Compare(api.MustParse("0.7")) < 0 {
+					exportFlags = append(exportFlags, []string{"-run-image", exportRegFixtures.ReadOnlyRunImage}...)
+				}
+				exportArgs := append([]string{ctrPath(exporterPath)}, exportFlags...)
+				exportedImageName = "/tmp/oci-image-" + h.RandString(10)
+				exportArgs = append(exportArgs, exportedImageName)
+
+				output := h.DockerRun(t,
+					exportImage,
+					h.WithFlags(append(
+						dockerSocketMount,
+						"--env", "CNB_PLATFORM_API="+platformAPI,
+						"--env", "CNB_REGISTRY_AUTH="+exportRegAuthConfig,
+						"--network", exportRegNetwork,
+					)...),
+					h.WithArgs(exportArgs...),
+				)
+				h.AssertStringContains(t, output, "Saving /layers/image")
 			})
 		})
 	}
